converted hcf_util to hcf_snap_reporter to use HeraCorrelator as base class
