+++ control_software/scripts/hera_snap_redis_monitor.py
@@ -7,8 +7,8 @@ import argparse
 from os import path as op
 from datetime import datetime
 
-from hera_corr_f import HeraCorrelator, __version__, __package__
-from hera_corr_f import hcf_util as hcfu
+from hera_corr_f import __version__, __package__
+from hera_corr_f import hcf_snap_reporter
 from hera_corr_cm.handlers import add_default_log_handlers
 
 logger = add_default_log_handlers(logging.getLogger(__file__))
@@ -58,8 +58,10 @@ if __name__ == "__main__":
         for handler in logger.handlers:
             handler.setLevel(getattr(logging, args.loglevel))
 
-    corr = HeraCorrelator(redishost=args.redishost, redis_transport=(not args.noredistapcp),
-                          block_monitoring=False)
+    corr = hcf_snap_reporter.SnapReporter(redishost=args.redishost,
+                                          redis_transport=(not args.noredistapcp),
+                                          block_monitoring=False,
+                                          logger=logger)
     upload_time = corr.r.hget('snap_configuration', 'upload_time')
     print_ant_log_messages(corr)
 
@@ -85,8 +87,10 @@ if __name__ == "__main__":
         if corr.r.hget('snap_configuration', 'upload_time') != upload_time:
             upload_time = corr.r.hget('snap_configuration', 'upload_time')
             logger.info('New configuration detected. Reinitializing fengine list')
-            corr = HeraCorrelator(redishost=args.redishost, use_redis=(not args.noredistapcp),
-                                  block_monitoring=False)
+            corr = hcf_snap_reporter.SnapReporter(redishost=args.redishost,
+                                                  redis_transport=(not args.noredistapcp),
+                                                  block_monitoring=False,
+                                                  logger=logger)
             print_ant_log_messages(corr)
 
         # Recompute the hookup every time. It's fast
@@ -112,7 +116,7 @@ if __name__ == "__main__":
                     continue
                 if stream % 2 == 0:
                     # adc_stats returns both polarizations, so only read every other
-                    adc_stats = hcfu.get_adc_stats(corr, host, stream, logger)
+                    adc_stats = corr.get_adc_stats(corr, host, stream, logger)
                     this_pol = 'x'
                 else:
                     this_pol = 'y'
@@ -120,11 +124,11 @@ if __name__ == "__main__":
                 # Build snap_rf_stats and add to redis
                 snap_rf_stats = {}
                 snap_rf_stats["timestamp"] = datetime.now().isoformat()
-                snap_rf_stats["autocorrelation"] = hcfu.get_auto(corr, host, stream, logger)
-                snap_rf_stats["eq_coeffs"] = hcfu.get_eq_coeff(corr, host, stream, logger)
-                snap_rf_stats["fft_of"] = hcfu.get_fft_of(corr, host, stream, logger)
+                snap_rf_stats["autocorrelation"] = corr.get_auto(corr, host, stream, logger)
+                snap_rf_stats["eq_coeffs"] = corr.get_eq_coeff(corr, host, stream, logger)
+                snap_rf_stats["fft_of"] = corr.get_fft_of(corr, host, stream, logger)
                 snap_rf_stats["histogram"] = adc_stats[this_pol]['histogram']
-                snap_rf_stats = hcfu.validate_redis_dict(snap_rf_stats)
+                snap_rf_stats = corr.validate_redis_dict(snap_rf_stats)
 
                 snaprf_status_redis_key = "status:snaprf:{}:{}".format(host, stream)
                 corr.r.hmset(snaprf_status_redis_key, snap_rf_stats)
@@ -142,9 +146,9 @@ if __name__ == "__main__":
                     ant_status[val] = snap_rf_stats[val]
                 for val in ['adc_mean', 'adc_power', 'adc_rms', 'histogram']:
                     ant_status[val] = adc_stats[this_pol][val.split('_')[-1]]
-                ant_status.update(hcfu.get_pam_stats(corr, host, stream, pol, logger))
-                ant_status.update(hcfu.get_fem_stats(corr, host, stream, logger))
-                ant_status = hcfu.validate_redis_dict(ant_status)
+                ant_status.update(corr.get_pam_stats(corr, host, stream, pol, logger))
+                ant_status.update(corr.get_fem_stats(corr, host, stream, logger))
+                ant_status = corr.validate_redis_dict(ant_status)
                 ant_status_redis_key = "status:ant:{:d}:{:s}".format(ant, pol)
                 corr.r.hmset(ant_status_redis_key, ant_status)
